Deep Review WIP - bots/instagram
- instagram specific actions, helpers, and constants were cleaned up,
  consolidated, and comments improved
- old helper methods for running async conditional functions were moved
  into actions since they are considered BotAction's, Conditional*
