--- src/botmation/bots/instagram/actions/feed.ts
@@ -1,19 +1,0 @@
-import { Page } from 'puppeteer'
-
-import { forAsLong } from '../../../actions/utilities'
-
-/**
- * TODO: implement, use doWhile(scrolling feed)(like posts) until we can't scroll anymore (NO, found bottom of page, that message that there are no more new posts)
- * @description   Favorite all published photos from these usernames
- * @param usernames 
- */
-// export const favoriteAllFrom = (...usernames: string[]): BotAction<void> => async(page: Page) => 
-  // while (not end of content) { 
-  //    consume post data (author name), like button & class for isLike status then "like" the content if a username 
-  //    finally scroll (may be optional, bot does not need a window since scroll can occur automatically upon press of "like" button)
-  // }
-
-  // await forAsLong()(
-
-  // )
-  // (page)
