+++ src/botmation/bots/instagram/constants/urls.ts
@@ -1,5 +1,6 @@
 /**
- * @description  Exported constants representing URL's belonging to Instagram
+ * @description  Crawled URL's belonging to Instagram
+ *               Broken apart into BASE & EXT (URL extensions)
  */
 export const INSTAGRAM_BASE_URL = 'https://www.instagram.com'
-export const INSTAGRAM_URL_EXT_LOGIN = 'accounts/login'
+export const INSTAGRAM_URL_EXT_LOGIN = '/accounts/login'
